+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Dev
 *(2022-07-08)*
 
 * Added the option to fix the order of tasks given to the workers (`#46`_)
+* Fixed a bug where updated WorkerPool parameters aren't used in subsequent ``map`` calls when ``keep_alive`` is enabled
 
 .. _#46: https://github.com/Slimmer-AI/mpire/issues/46