+++ docs/changelog.rst
@@ -1,10 +1,10 @@
 Changelog
 =========
 
-Dev
----
+2.5.0
+-----
 
-*(2022-07-08)*
+*(2022-07-25)*
 
 * Added the option to fix the order of tasks given to the workers (`#46`_)
 * Fixed a bug where updated WorkerPool parameters aren't used in subsequent ``map`` calls when ``keep_alive`` is enabled